+++ app/javascript/mastodon/components/area_header.tsx
@@ -1,6 +1,5 @@
 import type { Account } from 'mastodon/models/account';
-
-import { getAreaEngName, getAreaShortName } from './area';
+import { getAreaEngName, getAreaShortName } from 'mastodon/utils/area';
 
 interface Props {
   account: Account;
