+++ app/javascript/mastodon/utils/area.ts
